+++ CHANGELOG.md
@@ -18,9 +18,11 @@ The types of changes are:
 ## [Unreleased](https://github.com/ethyca/fides/compare/2.9.1...main)
 
 ### Added
+* Allow users to configure their username and password via the config file [#2884](https://github.com/ethyca/fides/pull/2884)
 * Add an Organization Management page (beta) [#2908](https://github.com/ethyca/fides/pull/2908) 
 
 ### Changed
+
 * Improved standard layout for large width screens and polished misc. pages [#2869](https://github.com/ethyca/fides/pull/2869)
 * Changed UI paths in the admin-ui [#2869](https://github.com/ethyca/fides/pull/2892)
   * `/add-systems/new` --> `/add-systems/manual`
@@ -31,6 +33,7 @@ The types of changes are:
 * Polish Admin UI header & navigation [#2897](https://github.com/ethyca/fides/pull/2897)
 
 ### Fixed
+
 * Restricted Contributors from being able to create Owners [#2888](https://github.com/ethyca/fides/pull/2888)
 * Allow for dynamic aspect ratio for logo on Privacy Center 404 [#2895](https://github.com/ethyca/fides/pull/2895)
 * Allow for dynamic aspect ratio for logo on consent page [#2895](https://github.com/ethyca/fides/pull/2895)
@@ -38,6 +41,7 @@ The types of changes are:
 ## [2.9.1](https://github.com/ethyca/fides/compare/2.9.0...2.9.1)
 
 ### Added
+
 * Added Attentive erasure email connector [#2782](https://github.com/ethyca/fides/pull/2782)
 
 ### Changed